+++ src/test/java/cc/kostic/gec/primitives/ExpirationTest.java
@@ -4,6 +4,7 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import java.math.BigDecimal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.*;
@@ -31,7 +32,6 @@ public class ExpirationTest {
 			long unixmS = date.getTime();
 			System.out.println("Unix Timestamp for " + exp + ": " + unixmS);
 			
-			
 		} catch (ParseException e) {
 			throw new RuntimeException(e);
 		}
@@ -45,7 +45,7 @@ public class ExpirationTest {
 	public void testIks() {
 		Instant instant = Instant.ofEpochMilli(1752103030606L);
 		// DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");		// ISO date
-		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dMMMyy");					// 9JUL20 bez vodece nule
+		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dMMMyy");						// 9JUL20 bez vodece nule
 		LocalDateTime localDateTime = LocalDateTime.ofInstant(instant, ZoneOffset.UTC);
 		String fs1 = localDateTime.format(formatter).toUpperCase();	// deribit voli uppercase
 		System.out.println(fs1);
@@ -54,13 +54,13 @@ public class ExpirationTest {
 	
 	@Test
 	public void testGetExpiration_timestamp() {
-		Expiration exp27mar26 = new Expiration(1774598400000L);		// Deribit expiration "27MAR26" = Fri Mar 27 2026 08:00 GMT
-		assertEquals(exp27mar26.getExpiration_timestamp(), 1774598400000L, "sam sebe");
+		Expiration exp27mar26 = new Expiration(BigDecimal.valueOf(1774598400000L));				// Deribit expiration "27MAR26" = Fri Mar 27 2026 08:00 GMT
+		assertEquals(exp27mar26.getExpiration_timestamp(), BigDecimal.valueOf(1774598400000L), "sam sebe");
 		assertEquals(exp27mar26.getExpirationStringISO(), "2026-03-27T08:00:00Z", "iso date");
 		assertEquals(exp27mar26.getExpirationShortFmt(), "27MAR26", "deribit kratki format");
 
-		Expiration bezVodeceNule = new Expiration(1772953200000L);		// Sun Mar 08 2026 07:00:00 GMT+0000
-		Expiration bezVodeceNulePlus = new Expiration(1772953200001L);	// plus 1mS ali je u istom danu
+		Expiration bezVodeceNule = new Expiration(BigDecimal.valueOf(1772953200000L));			// Sun Mar 08 2026 07:00:00 GMT+0000
+		Expiration bezVodeceNulePlus = new Expiration(BigDecimal.valueOf(1772953200001L));		// plus 1mS ali je u istom danu
 		assertEquals(bezVodeceNule.getExpirationShortFmt(), "8MAR26", "deribit kratki format - bez vodece nule");
 		assertEquals(bezVodeceNulePlus.getExpirationShortFmt(), "8MAR26", "mora biti u okviru istog datuma");
 	}
@@ -79,10 +79,10 @@ public class ExpirationTest {
 	
 	@Test
 	public void testCompareTo() {
-		Expiration a = new Expiration(30);
-		Expiration hi = new Expiration(40);
-		Expiration lo = new Expiration(20);
-		Expiration eq = new Expiration(30);
+		Expiration a = new Expiration(BigDecimal.valueOf(30L));
+		Expiration hi = new Expiration(BigDecimal.valueOf(40L));
+		Expiration lo = new Expiration(BigDecimal.valueOf(20L));
+		Expiration eq = new Expiration(BigDecimal.valueOf(30L));
 		assertTrue(a.compareTo(a)==0, "sam sa sobom");
 		assertTrue(a.compareTo(lo)>0, "ja sam veci");
 		assertTrue(a.compareTo(hi)<0, "ja sam manji");
